NULL conf pointer after free
Fixes an issue where if we have already seen the topic we may have created a new rd_kafka_topic_conf_t object
that will leak memory. If we have already seen the topic, we can check that the rd_kafka_topic_conf_t
pointer is not NULL and free it ourselves, otherwise if it is NULL we know that the internals have already
deallocated conf. This commit mainly prevents a double-free error.