Fix minor indentation issue
A single line in that function was indented with tabs insteads of
spaces, which made it appear as it was part of an else clause even
though it wasn't (a "goto fail" kind of thing). At least some version(s)
of GCC 6 complained about such differences in whitespace and it was
misleading anyway, so adjust the whitespace accordingly.