Count partially sent buffers as in-flight when timeout scanning (issue #756)
This makes sure the connection is closed when in-flight requests time out.
Failure to do so would cause an invalid packet stream since the timed
out buffer would be removed and the next one sent which will then occur
in the middle of the partially sent buffer, thus eventually giving
a parse error (InvalidReceiveException) on the broker.

Big thanks to @andrey-sra for tracking down this issue.