Fix undefined references to static const ints (#961, #962, @bbfgelman1)
This patch fixes an issue when compiling code with librdkafka
with all optimizations disabled. The library declares a number
of static constant integer values in headers, but not in
compiled files. This causes any code that passes these constants
directly to a function accepting a const reference to result in
undefined reference errors when linking.

For RdKafka::Producer, change the constants from static const int
to being members of an anonymous enum. They can be referred to
by the same name, but now behave more like literals instead of
static member variables.

For RdKafka::Topic, declared the constants and their values
in the cpp file. As the values now appear only in the cpp file,
use the C macros instead of duplicating the values.

The values have to appear in the cpp file instead of the header
because of a VS2013 bug:

http://stackoverflow.com/questions/31074546/vs2013-static-const-already-defined
(cherry picked from commit a0d19638564de8ff14e7c2c7b0ac81a54c28f69e)