Partial d3676c5cdce28 revert: Don't emit TIMED_OUT_QUEUE
While it might be interesting for librdkafka developers
to see the difference between in-queue and in-flight timeouts
this change also changed the expected error returns from delivery reports, et.al.,
and thus was a badly documented API:ish change.

This commit reverts back to the previous functionality with emitting TIMED_OUT
for in-queue timeouts.