posix header: prepare support for z/os
z/Os is somewhat posix. but the compiler does not support
stuff like likely()/unlikely() or some fancier __attribute__(())

More importantly xcl does not support the __tread annotation.

We will have to work around that later.. and RD_TLS remains undefined
for z/OS

z/OS c library has a usleep function that can do micro-second based sleep
with the caveat that it cannot be called for a sleep superior or equal
to 10000000 micro-sec. (iow 1s).