Interceptors are no longer copied by rd_kafka_conf_dup()
Silently copying registered interceptors leads to multiple references
to the same interceptor instance data which the interceptor does not
know about.

The new behaviour copies no interceptors in rd_kafka_conf_dup()
but relies on the copying of `plugins.library.path` to trigger
the conf_init() of plugin interceptors.

It is also possible for an interceptor to register an on_conf_dup()
method which is called on the new configuration object prior to
copying the configuration properties.

As a safety precaution, duplicate interceptors (same ic_name and method)
will be rejected. The only exception is on_conf_destroy().

This also adds the final configuration object as a const parameter
to the on_new() interceptor method so it can make use of/inspect the final
configuration.