Proper re-sends of ProduceRequests in-transit on connection down (#1390)
When the broker connection is torn down, any outstanding ProduceRequests
will have their associated messages put back on their partition queues,
which will trigger future ProduceRequests, but the original ProduceRequest
is still retried with the serialized messages, but without the message
association. This results in duplicate messages, or error logs.