Avoid allocating BIOs and copying for base64 processing.
OpenSSL offers a more direct API for base64 encoding and decoding.
Previously the code created a couple of BIOs and used the streaming
interface, this results in unneccessary allocations.

Also, we've received a request to integrate this library with TensorFlow
and the base64 BIO interface is one of the things we're trying to trim
from our version of OpenSSL.

Note: if there's an easy alternative to using OpenSSL for base64
processing then I would recommand that. OpenSSL has some odd behaviours
around ignoring whitespace etc. While they don't immediately appear to
matter here, something simplier would be preferable.