dynamically initialize rd_kafka_err_descs on solaris
There is a bug in the way Sun Studio generates relocations for pointers
into the middle of strings.  That means the name field of all the
rd_kafka_err_desc objects ends up being null instead of the correct
name.  Basically you can't have a compile time initialized pointer to
any part of a string literal other than the beginning.  Since the layout
of rd_kafka_err_desc is part of the ABI and we can't compile time
initialize it on Solaris we need to initialize it as part of the global
library initialization.