Don't let metadata ERR_UNKNOWN set topic state to non-existent
The old code used ERR_UNKNOWN to detect auto topic creation failures,
but I could not find a broker version from >= 0.8.2.1 that returned
that error in that case.
ERR_UNKNOWN may be temporarily returned in metadata as the topic is
being created in which case we should just re-query metadata until
there is a proper error or the topic creation is successful.