Fix crash: failed ops enq could be handled by original dest queue callbacks
If an rko was enqueued on a forwarded queue who's final destination
queue was disabled (but not the initial one), the rko would be set up
with the orignal queue's rko_serve callback which would then be
triggered on the reply queue with the rko type changed to a REPLY
with an error set, something these callbacks will not gracefully
handle (and shouldn't need to).

This fix makes sure the original dest queue rko_serve callback is only
set on the rko if it is going to be successfully enqueued.