interceptor_test: don't link any libraries
The example plugin library now will not link to librdkafka
to avoid the case where the librdkafka soname or path has been
changed.
Instead we rely on the parent process to have properly linked
librdkafka and its dependencies, which should always be the case
since the plugin library is loaded from librdkafka configuration.