max.poll.interval.ms: properly handle blocking calls
Previous to this fix the app_polled time was only updated upon
entry to a (possibly) blocking function, such as consumer_poll().
If the call had a timeout > max.poll.interval.ms and there were no messages
for at least max.poll.interval.ms, the max poll check would kick in and
evict the consumer from the group due to inactivity.

We now signal to app_polled that we're in a blocking call
which prevents the max poll check to fire.