Fail configuration when requested library isn't present
If the user specifies --enable-XXX, fail configuration if XXX isn't
present. The prior behavior required the user to manually inspect the
configure output to determine whether the feature was actually enabled.
This was particularly problematic with downstream build systems, like
Cargo with rust-rdkafka, where the configure output is hidden from users
if configuration is successful.

The default behavior remains unchanged; if a user does not explicitly
request a feature with --enable-XXX nor dis-request it with
--disable-XXX, and the feature is enabled by default, then it is not be
an error if the feature cannot be found.

Note that this commit additionally excludes the `ENABLE_*` options from
the confugration summary, as they were misleading. For example, if a
user passed --enable-zstd, but zstd was not found on the system, the
configuration summary would report "ENABLE_ZSTD y"--merely reporting
that the user had requested zstd support, not that zstd support would
actually be available in the built library. (If the user does desire
that information, they can find it in the BUILT_WITH line, which
includes only the features that were actually possible to enable.)