Shave another 1s off cgrp initialization
Querying for the cgrp coordinator does not need to back off when there
are no brokers connected, as no requests were sent. This avoids an
accident of timing that would cause cgrp initialization to wait an extra
turn of the main rdkafka thread loop, which would add an unnecessary
second of latency.

Inspired by c64b652689.