Broker socket would be closed twice if connect() returned error
This would really only happen if a socket_cb was set that introduced errors,
like sockem does. The result was that if another thread would create a new
socket between the two close calls the fd could be reused and spontaneously
closed by the second close call.
We've seen this in flaky sockem tests.