mklove: bundle all static libraries into librdkafka-static.a, regardless of --enable-static
Previously, any static library dependencies would be added to pkg-config's like
Libs.private: /path/to/libdependency.a

These dependencies are highly local and their use in a pkg-config file provide
more harm than good, especially when librdkafka is built with --install-deps
that builds and links a temporary static library of the dependency.

The fix here is to always bundle these static library dependencies inside
librdkafka-static.a.