Assignor: de-conflate owned/assigned partitions (#3306)
rkgm_assignment, which is the member assignment after running the assignor,
was mixed up with rkgm_owned, which is the current member assignment before
running the assignor. This resulted in the sticky assignor not taking the
current assignment into consideration on rebalance and thus not being able to
provide the stickyness.