Correctly handle empty aborted transactions
Seems like there is a case where aborted transactions will have an
ABORT ctrl message marker, but not be included in the aborted transactions
list.
This would cause the next aborted-transaction-offset to be popped from the
aborted transaction list even though it did not match the aborted transaction
marker. This in turn could lead to aborted messages for subsequent aborted
transactions in the same MessageSet to not be skipped and instead delivered
to the application.

The fix is to silently (unless debug) ignore these unsolicited abort markers
if they're not in the aborted transaction list and not pop off the next
non-matching offset from the aborted transaction list.