Fix clang-cl on windows compilation
Clang when run as clang-cl on windows has subtle issues with SIMD intrinsics. Essentially it tries to parse its own headers but gets confused about the handling of none standard `__attribute__` markers. This leads it to confuse the meaning of certain SIMD types.

Fortunately librdkafka (and snappy as embedded) only need some basic operations that clang (even as clang-cl` support). As such we can use the builtins