Always provide a librdkafka-static.a
Even if there are no dependency .a to bundle we should still provide a librdkafka-static.a, since it
may be sufficient anyway.

This change originally stems from the RPM package librdkafka.spec including librdkafka-static.a, but
depending on Redhat distro it may or may not need static library dependencies, which in turn would
sometime not create a librdkafka-static.a, thus failing RPM packaging.