Allow overriding DNS resolution
Introduce a new `resolve_cb` configuration parameter for controlling DNS
resolution. When provided, librdkafka invokes the callback when
resolving the address of a Kafka broker instead of calling `getaddrinfo`
directly.  The callback is expected to function exactly as
`getaddrinfo`, except that it is additional provided with the `opaque`
value.

The callback will enable tunneling to a Kafka broker (e.g., via SSH or
AWS PrivateLink). The broker addresses returned by the bootstrap
protocol will not be directly routable from the client; the address
resolution callback allows the end user to rewrite the broker addresses
as appropriate for the tunnel.