Add more explicit .clang-format files and warn if clang-format version != 10
Using the --dump-files option on clang-format-10, this guards against
any changes to the defaults, or changes to the 'BasedOnStyle' styles.
For .clang-format-cpp, because of the way we are using it in the script
by converting it to the key/value format, a number of options from the
--dump-files option has to be omitted, hence why we need to retain
the BasedOnStyle. For .clang-format, we no longer need this.

Additionally, style-format.sh warns in case clang-format version is
not 10.x.y.