Add a random key mode
Currently rdkafka_performance offers a mode where you use
a fixed key (-kfoobar) or no key at all, in which case the
partitioner logic will choose the partition based on its
own logic.

It is also useful to have a random key mode: this tends to
spread messages evently across partitions (and hence brokers)
which can have quite differnt characteristics as compared
to the no-key partitioner behavior.

This change adds the -KN mode which uses a random key of
length N, with a new key for each message.