Don't reset max.poll.interval.ms limit on every poll (#4176)
Earlier on, we changed the code to prevent max.poll.interval.ms
from being triggered in case we were inside librdkafka in any
sort of a poll call. Top achieve this, blocked the timer using
rd_kafka_app_poll_blocking, and reset it at the end of the call.

This doesn't work correctly in the cases where we're simply
polling an unrelated queue, like the log queue. In that case,
just by polling the log queue, the timer is reset, despite us
not actually consuming anything (or doing any consume poll).

At the same time, it's a reasonable expectation that
max.poll.interval.ms won't be triggered while we are
doing any sort of consumer poll.

This commit takes care of both the cases. 
The methods (from public API) which block/reset the timer are:

    rd_kafka_consume_batch
    rd_kafka_consume_batch_queue
    rd_kafka_consume_callback
    rd_kafka_consume_callback_queue
    rd_kafka_consume
    rd_kafka_consumer_poll