Adds config property for overriding TLS hostname
Fixes #3321

This diff adds a new config property `ssl.endpoint.hostname` that can be
used to override the hostname users expect to be present in TLS certs/SANs
when `ssl.endpoint.identification.algorithm` is set to `https`.

This is useful in cases where you have a CA that issues certs that
contain names/SANs that do not correspond to the broker hostname.

One benefit of using a config property to do this is that it the
functionality is instantly available in all language bindings.

Note that one can accomplish this by setting a custom `cert_verify_cb`,
however the barrier to entry is pretty high for language binding users
as they'd need to deal w/ FFI. In my experience it is more common to
disable hostname verification instead.