Fix linting and Makefile issues (#2)
* Add single-offset-storage method for performance

In promisified mode, offsets are stored after every single message
Calling into FFI is already not cheap and on top of that, the
offsetsStore method does a lot of allocations because it needs to make a
vector etc. etc. This method change yields a benefit of 8%.

Also reduces dynamic casts by storing a reference to the rdkafka
consumer.

* Remove jshint and expand eslint coverage

Many changes within, but only cosmetic:
1. remove jshint as it's rather unmaintained and we don't need 2
   linters.
2. update eslint to latest possible version.
3. update eslint coverage to include old API and promisified tests.
4. update makefile accordingly
5. fix lint errors.

* Update Makefule, update cpplint to latest version

* Comply with cpplint