Handle the case where current group coordinator is decommissioned without leaving dangling references until the coordinator is changed
Test 0151 fix. Given the find coordinator response adds a new broker (not a logical one, a learned one to set into `rkcg_curr_coord`)
Removed brokers can be added again even if not present in metadata. This is mock cluster only problem as in a
real cluster a broker that is set down cannot be a coordinator. This commit changes the coordinator before
setting down a broker that is current coordinator