Purge brokers no longer reported in metadata (#4557)
* Fix for brokers with different Ids but same host:port

The Kafka protocol allows for brokers to have multiple host:port pairs
for a given node Id, e.g. see UpdateMetadata request which contains a
live_brokers list where each broker Id has a list of host:port pairs. It
follows from this that the thing that uniquely identifies a broker is
its Id, and not the host:port.

The behaviour right now is that if we have multiple brokers with the
same host:port but different Ids, the first broker in the list will be
updated to have the Id of whatever broker we're looking at as we iterate
through the brokers in the Metadata response in
rd_kafka_parse_Metadata0(), e.g.

 Step 1. Broker[0] = Metadata.brokers[0]
 Step 2. Broker[0] = Metadata.brokers[1]
 Step 3. Broker[0] = Metadata.brokers[2]

A typical situation where brokers have the same host:port pair but
differ in their Id is if the brokers are behind a load balancer.

The NODE_UPDATE mechanism responsible for this was originally added in
b09ff60c ("Handle broker name and nodeid updates (issue #343)") as a way
to forcibly update a broker hostname if an Id is reused with a new host
after the original one was decommissioned. But this isn't how the Java
Kafka client works, so let's use the Metadata response as the source of
truth instead of updating brokers if we can only match by their
host:port.

* Fix for purging brokers no longer reported in metadata

Brokers that are not in the metadata should be purged from the internal
client lists. This helps to avoid annoying "No route to host" and other
connection failure messages.

Fixes #238.

* Remove the possibility to modify rkb_nodeid after rkb creation.

* Remove locking when accessing rkb_nodeid as it's now set only on creation and not modified anymore

* Add new brokers and reassign partitions in the mock cluster

* Remove bootstrap broker after receiving learned
ones. Wait decommissioned threads after they've stopped instead of on termination.

* Handle the _DESTROY_BROKER local error,
triggered when a broker is removed
without terminating the client.

* Test 0151 improved with cluster replacement
and cluster roll

* Fix for test 0105, do_test_txn_broker_down_in_txn:
remove left references when decommissioning a broker
and avoid it's selected as leader again or that partitions
are delegated to it

* Avoid selecting a configured broker as a logical or telemetry broker

* Avoid selecting terminating brokers for sending calls or new connections

* Remove addressless count and avoid counting the logical
broker for the all brokers down error, to send
the error in all the cases

* Test: verify that decommissining a broker while adding a new one with same id isn't causing problems

* Handle the case where current group coordinator
is decommissioned without leaving dangling references
until the coordinator is changed

Test 0151 fix. Given the find coordinator response adds a new broker (not a logical one, a learned one to set into `rkcg_curr_coord`)
Removed brokers can be added again even if not present in metadata. This is mock cluster only problem as in a
real cluster a broker that is set down cannot be a coordinator. This commit changes the coordinator before
setting down a broker that is current coordinator

* Remove the decommissioning broker from rk_broker_by_id when starting to avoid multiple
instances are added to the list with same id.
The decommissioned broker returned by the find can lead to multiple brokers with same id being added.

* Don't select logical brokers at all
for general purpose request like metadata ones

* Schedule an immediate connection when there are no brokers connecting nor requests for connection.
When we're in this state, if we respect the sparse connection interval, there's no event that notifies the awaiters at `rd_kafka_brokers_wait_state_change` given it's an interval and not a timer.
This is more visible when brokers are decommissioned and there's no broker down even causing the notification.
Check it with test `0113` subtest `u_multiple_subscription_changes`.

* Remove all configured brokers when there are learned
ones. This is to avoid leaving connections to the boostrap brokers that are continued to be used instead of the learned one, adding additional requests that can be later purged by the decommissioning of that last configured broker.

* Change test 0075 after removing all bootstrap brokers.
This will be reverted with KIP-899

* Remove rk_logical_broker_up_cnt

* [test 0151] Simplify the test removing `await_verification`. It's possible to await for the correct list of brokers in all tests given since decommissioning brokers are excluded from that list

* Remove broker state from labels

* Remove `nodeid` from op

* Use `rk_broker_by_id` for learned broker ids to return sorted broker ids

* Verify nodename change through a
test log interceptor.
Used the test log interceptor for test 0151 too


---------

Co-authored-by: Emanuele Sabellico <esabellico@confluent.io>