[KIP-848] Handle topic authorization failure on consumer group heartbeat (#5009)
Handle topic authorization failure on consumer
group heartbeat, in this case the heartbeat is retried
with given interval or with exponential backoff
if it's the first heartbeat. An error is returned
to the user in this case and isn't sent again for the given heartbeat interval or at least 5 seconds.