Change test expected behaviour in case of out of order sequence error when it's not the first error.
In such cases all messages can be retried because if the first error was persistent it had caused a epoch bump already and if it was possibly persisted it can be retried together with next sequences.

If it bumps the epoch, instead, there's a chance it causes side effects as message duplications on other partitions.