Fix for a loop of re-bootstrap sequences (#5086)
in case the client reaches the all brokers down state.

The client continue to select the bootstrap brokers given they have no connection and doesn't re-connect to the learned ones. Fixed by giving a chance to connect to the learned brokers even if there are new ones that never tried to connect.

Closes #5088